drm: amd: avoid 'static const' variables in headers
The most common warning in a 'make W=1' build is

drivers/gpu/drm/amd/amdgpu/../display/include/fixed31_32.h:72:32: error: 'dc_fixpt_pi' defined but not used [-Werror=unused-const-variable=]

We get that 279 times, along with thousands of other warnigns
about constant variables that are declared in amdgpu header files.

The way that the amd driver does it is a bit unconventional, but
not strictly wrong.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>