[SUBMITTED 20190305] EDAC: i10nm, skx: fix randconfig builds
In a configuration where one of the two drivers is built-in and the
other one is a module, kbuild tries to add the modular file into
vmlinux, and fails with

drivers/edac/skx_common.o:(.rodata+0xc0): undefined reference to `__this_module'

We either have to make both drivers be configured the same way all the
time, or make skx_common a separate module.

This takes the second approach, which is more logical, but has the downside
of requiring lots of new EXPORT_SYMBOL_GPL() statements.

Fixes: d4dc89d069aa ("EDAC, i10nm: Add a driver for Intel 10nm server processors")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>