ARM: uaccess: reimplement access_ok() in C
It turns out that modern compilers can do better than the
inline assembly because they are able to combine repeated
checks of the same area and schedule the instructions.

The adds/sbcscc pair of instructions is replaced with two
comparisons here, but we are able to use the resulting
condition code directly rather than having to go through
a register.

Note: the return value of __range_ok() is inverted now,
it returns '1' if the access is ok rather than '0'.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>