[SUBMITTED 20190325] netfilter: nat: avoid unused-variable warning
masq_refcnt6 was added at the start of the file, but it is
only used in the option IPv6 section of the file, causing
a harmless compiler warning if IPv6 is disabled:

net/netfilter/nf_nat_masquerade.c:15:21: error: 'masq_refcnt6' defined but not used [-Werror=unused-variable]
 static unsigned int masq_refcnt6 __read_mostly;

Move the variable next to the user to avoid that warning.

Fixes: 46f7487e161b ("netfilter: nat: don't register device notifier twice")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
----
This is an older patch stack, for some reason I seem to have never
sent it, and I can't find any indication of anyone else sending
a similar fix, so sending this out now.

If it's already fixed upstream, please ignore.