[alternative SUBMITTED 20190322] bnxt: avoid uninitialized variable use
clang found a case in which an uninitialized variable is
being used in the rx error handling:

drivers/net/ethernet/broadcom/bnxt/bnxt.c:1612:6: error: variable 'len' is used uninitialized whenever 'if' condition
      is true [-Werror,-Wsometimes-uninitialized]
        if (rxcmp1->rx_cmp_cfa_code_errors_v2 & RX_CMP_L2_ERRORS) {
            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
drivers/net/ethernet/broadcom/bnxt/bnxt.c:1703:19: note: uninitialized use occurs here
        cpr->rx_bytes += len;
                         ^~~
drivers/net/ethernet/broadcom/bnxt/bnxt.c:1612:2: note: remove the 'if' if its condition is always false
        if (rxcmp1->rx_cmp_cfa_code_errors_v2 & RX_CMP_L2_ERRORS) {
        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
drivers/net/ethernet/broadcom/bnxt/bnxt.c:1540:18: note: initialize the variable 'len' to silence this warning
        unsigned int len;
                        ^
                         = 0

There was a previous fix for the same issue when coverity found it,
but that patch only caught two of the three conditions leading to this.

I'm not sure what the correct behavior here should be, either we
intialize the length earlier, or we don't use it and skip that
step as in the other patch. This picks the first approach based
on my best guess.

Fixes: e7e70fa6784b ("bnxt_en: don't update cpr->rx_bytes with uninitialized length len")
Cc: Colin Ian King <colin.king@canonical.com>
Cc: Michael Chan <michael.chan@broadcom.com>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>