[SUBMITTED 20190322] tipc: avoid a clang -Wuninitialized warning
clang notices that we pass 'maddr' into the tipc_bearer_xmit() function
without having initialized it first:

net/tipc/node.c:831:6: error: variable 'maddr' is used uninitialized whenever 'if' condition is false
      [-Werror,-Wsometimes-uninitialized]
        if (!tipc_link_is_establishing(l)) {
            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~
net/tipc/node.c:847:46: note: uninitialized use occurs here
        tipc_bearer_xmit(n->net, bearer_id, &xmitq, maddr);
                                                    ^~~~~
net/tipc/node.c:831:2: note: remove the 'if' if its condition is always true
        if (!tipc_link_is_establishing(l)) {
        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
net/tipc/node.c:821:31: note: initialize the variable 'maddr' to silence this warning
        struct tipc_media_addr *maddr;
                                     ^
                                      = NULL

This is harmless because it won't use 'maddr' if the
queue is empty, but it's better to make that explicit
and not even call the function in that case.

As clang is able to inline the check, it then notices that
the code is safe.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>