[SUBMITTED 20190722] [RESEND] mfd: davinci_voicecodec: remove pointless #include
When building davinci as multiplatform, we get a build error
in this file:

drivers/mfd/davinci_voicecodec.c:22:10: fatal error: 'mach/hardware.h' file not found

The header is only used to access the io_v2p() macro, but the
result is already known because that comes from the resource
a little bit earlier.

Acked-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
I sent this in June, but it did not get picked up into the mfd
tree. Sending the same one again in case it got lost.