[SUBMITTED 20190709] [v2] ARM: mtd-xip: work around clang/llvm bug
llvm gets confused by inline asm with .rep directives, which
can lead to miscalculating the number of instructions inside it,
and in turn lead to an overflow for relative address calculation:

/tmp/cfi_cmdset_0002-539a47.s: Assembler messages:
/tmp/cfi_cmdset_0002-539a47.s:11288: Error: bad immediate value for offset (4100)
/tmp/cfi_cmdset_0002-539a47.s:11289: Error: bad immediate value for offset (4100)

This might be fixed in future clang versions, but is not hard
to work around by just replacing the .rep with a series of
eight unrolled nop instructions.

As Russell points out, the original code uses an undocumented
assembler directive, as .rep is normally spelled .rept, though
the shorter form is common on arch/x86 as well.

Link: https://bugs.llvm.org/show_bug.cgi?id=42539
Link: https://godbolt.org/z/DSM2Jy
Acked-by: Nick Desaulniers <ndesaulniers@google.com>
Reviewed-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
v2: mention .rep/.rept
    add missing "Link:" keyword.