[v3] blkdev: always export SECTOR_SHIFT
When CONFIG_BLOCK is disabled, SECTOR_SHIFT is unknown, and this leads
to a failure in the testing infrastructure added from commit
43c78d88036e ("kbuild: compile-test kernel headers to ensure they are
self-contained"):

In file included from <built-in>:3:
include/linux/iomap.h:76:48: error: use of undeclared identifier 'SECTOR_SHIFT'
        return (iomap->addr + pos - iomap->offset) >> SECTOR_SHIFT;

If we want to keep build testing all headers, the macro needs to
either be defined, or not used. Move it out of the #ifdef
section to ensure it is visible.

Fixes: db074436f421 ("iomap: move the direct IO code into a separate file")
Link: https://lore.kernel.org/lkml/20190718125509.775525-1-arnd@arndb.de/T/
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
v3: remove pointless #ifndef checks around the definition
    fix reference commit id
v2: move the macros outside of the #ifdef instead of adding another #ifdef
    around the user