ARM: iop13xx: remove mach/memory.h
After removing the omap1 __arch_dma_to_virt/__arch_virt_to_dma
implementation, I had a look at the other remaining ones.  In case
of iop13xx, it is quite tricky to find out what this is actually for,
as I could not find any driver code that actually relies on it.

iop13xx has a range of memory-mapped registers that are accessible
from both the CPU and DMA master capable devices, the registers on here
include serial, i2c, adma and others.

Passing an __iomem pointer corresponding to these mmio registers
into dma_map_single() triggers a special case that will return the
corresponding physical address.

However, this appears to be completely pointless, as we do not use slave
DMA on any of those devices, and nothing ever tries to map the registers
for DMA. My best guess is that this code was written to support a
potential use case that was never added to the kernel, so I propose
removing this one.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>