[RFC] iop: use GENERIC_IRQ_MULTI_HANDLER
This is an attempt to convert the interrupt handler for iop32x to
GENERIC_IRQ_MULTI_HANDLER, converting the get_irqnr_preamble and
set_irqnr_and_base into a C function for use with set_handle_irq().

The special handling for the cp6 interrupt registers makes this
a bit tricky. I tried to rewrite the arch_ret_to_user macro from
the header file into something that has a chance of working on
other architectures, but probably messed up the logic in the
process.

There is now a global flag 'iop32x_cp6_enabled' that tells us
whether cp6 got enabled earlier, and we check this flag while
returning to user space.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>